+++ carp_communication_package/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.5
+* `calendar` measure, probe, and datum added incl. default privacy protection 
+* `phone_log` measure redesigned as a periodic measure
+
 ## 0.3.4
 * upgrade to json_serializable v.2
 * privacy schema support for hashing text messages and phone numbers